♻️ refactor: split sandbox config into separate file
